Fix issues with loading games
Two issues encountered loading games-- the teams hash wasn't returned
in the teams method, making it impossible to actually load teams for
games.  The other issue is that failure to find a game.xml should
halt the process of making a game object, but previous code was
returning a broken game object instead.